refactor!: rename algorithms package to portfolio (#103)
E6.02 of the 2.0 refactor (breaking, no compat shim):
- fynance/algorithms/ -> fynance/portfolio/ (allocation + sizing).
- update all importers (backtest.cost, signal.mappers, models.rolling,
  top-level __init__) and tests.
- rename Sphinx pages algorithms*.rst -> portfolio*.rst, regenerate
  autosummary stubs, fix index/quickstart/cross-references.

BREAKING CHANGE: fynance.algorithms is removed; use fynance.portfolio.
