--- doc/source/algorithms.sizing.rst
@@ -1,14 +1,0 @@
-***************
-Position sizing
-***************
-
-Position sizing and transaction-cost primitives for realistic backtests: fractional Kelly (:func:`~fynance.algorithms.sizing.kelly_fraction`), volatility targeting (:func:`~fynance.algorithms.sizing.vol_target`) and turnover-based transaction costs (:func:`~fynance.algorithms.sizing.transaction_cost`).
-
-.. currentmodule:: fynance.algorithms.sizing
-
-.. autosummary::
-   :toctree: generated/
-
-   kelly_fraction
-   vol_target
-   transaction_cost
