+++ doc/source/models.rolling.rst
@@ -2,7 +2,7 @@
 Rolling models
 **************
 
-Tools and objects to roll a model over time-series data, for walk-forward evaluation. See also :mod:`fynance.algorithms.allocation` for rolling portfolio allocation algorithms.
+Tools and objects to roll a model over time-series data, for walk-forward evaluation. See also :mod:`fynance.portfolio.allocation` for rolling portfolio allocation algorithms.
 
 .. currentmodule:: fynance.models.rolling
