+++ doc/source/portfolio.sizing.rst
@@ -1,0 +1,14 @@
+***************
+Position sizing
+***************
+
+Position sizing and transaction-cost primitives for realistic backtests: fractional Kelly (:func:`~fynance.portfolio.sizing.kelly_fraction`), volatility targeting (:func:`~fynance.portfolio.sizing.vol_target`) and turnover-based transaction costs (:func:`~fynance.portfolio.sizing.transaction_cost`).
+
+.. currentmodule:: fynance.portfolio.sizing
+
+.. autosummary::
+   :toctree: generated/
+
+   kelly_fraction
+   vol_target
+   transaction_cost
