+++ fynance/__init__.py
@@ -21,7 +21,7 @@ analysis, portfolio allocation, and backtest trading strategies.
 
 Subpackages
 -----------
-algorithms      --- Financial algorithms
+portfolio       --- Portfolio allocation & sizing
 backtest        --- Backtest strategy tools
 estimator       --- Parameter estimation (Cython ARMA/GARCH)
 features        --- Features extraction
@@ -36,7 +36,7 @@ _wrappers   --- Fynance wrapper functions
 API stability policy (1.x series)
 ---------------------------------
 The symbols re-exported below from :mod:`fynance.models`,
-:mod:`fynance.algorithms.allocation`, :mod:`fynance.features` and
+:mod:`fynance.portfolio.allocation`, :mod:`fynance.features` and
 :mod:`fynance.estimator` form the **public, stable API** for the 1.x
 release line. Within 1.x:
 
@@ -64,16 +64,16 @@ __all__ = ['__version__']
 
 import sys as _sys
 
-from .algorithms import *
 from .backtest import *
 from .estimator import *
 from .features import *
 from .models import *
+from .portfolio import *
 
 # Aggregate each subpackage's public surface. Use ``sys.modules`` rather than the
 # package attributes, which a star import may have shadowed with a name that
 # collides with a submodule (e.g. the ``backtest`` engine function).
-for _name in ("models", "estimator", "features", "backtest", "algorithms"):
+for _name in ("models", "estimator", "features", "backtest", "portfolio"):
     __all__ += _sys.modules[f"{__name__}.{_name}"].__all__
 
 # Restore the subpackage attribute shadowed by such a collision so that
